requirements.txt
icecream==2.1.3
pendulum==2.1.2
PyYAML==6.0.1
tqdm==4.66.1